Add astropy registry support to datamodels and write support to nddata interface
The astropy registry interface provides a generic way to read and
write different astronomical images and datamodels has new functions
to support it, read, write, and identify. Also, setters have been to
the getters in the nddata interface. Most of the code to implement the
astropy registry and nddata nterface has been moved into a new file,
ndmodel.py. There is still some glue code in __init__.py,
model_base.py, and properties.py. DataModel now subclasses a new
class, NDModel, a mixin that implements the nddata interface.
