+++ jwst/datamodels/properties.py
@@ -6,7 +6,6 @@ from __future__ import absolute_import, division, unicode_literals, print_functi
 import copy
 import numpy as np
 import jsonschema
-import collections
 
 from astropy.extern import six
 from astropy.utils.compat.misc import override__dir__
@@ -185,6 +184,8 @@ class ObjectNode(Node):
             return self._instance == other
 
     def __getattr__(self, attr):
+        from . import ndmodel
+
         if attr.startswith('_'):
             raise AttributeError('No attribute {0}'.format(attr))
 
@@ -201,7 +202,7 @@ class ObjectNode(Node):
         if isinstance(val, dict):
             # Meta is special cased to support NDData interface
             if attr == 'meta':
-                node = MetaNode(val, schema, self._ctx)
+                node = ndmodel.MetaNode(val, schema, self._ctx)
             else:
                 node = ObjectNode(val, schema, self._ctx)
         elif isinstance(val, list):
@@ -407,93 +408,3 @@ def merge_tree(a, b):
     recurse(a, b)
     return a
 
-#---------------------------------------------
-# The following classes provide support
-# for the NDData interface to Datamodels
-#---------------------------------------------
-
-class MetaNode(ObjectNode, collections.MutableMapping):
-    """
-    NDData compatibility class for meta node
-    """
-    def __init__(self, instance, schema, ctx):
-        ObjectNode.__init__(self, instance, schema, ctx)
-
-    def _find(self, path):
-        if not path:
-            return self
-        
-        cursor = self._instance
-        schema = self._schema
-        for attr in path:
-            try:
-                cursor = cursor[attr]
-            except KeyError:
-                raise KeyError("'%s'" % '.'.join(path))
-            schema = _get_schema_for_property(schema, attr)
-            
-        return _make_node(cursor, schema, self._ctx)
-
-    def __delitem__(self, key):
-        path = key.split('.')
-        parent = self._find(path[:-1])
-        try:
-            parent.__delattr__(path[-1])
-        except KeyError:
-            raise KeyError("'%s'" % key)
-
-    def __getitem__(self, key):
-        path = key.split('.')
-        return self._find(path)
-    
-    def __iter__(self):
-        return MetaNodeIterator(self)
-
-    def __len__(self):
-        def recurse(val):
-            n = 0
-            for subval in six.itervalues(val):
-                if isinstance(subval, dict):
-                    n += recurse(subval)
-                else:
-                    n += 1
-            return n
-
-        return recurse(self._instance)
-
-    def __setitem__(self, key, value):
-        path = key.split('.')
-        parent = self._find(path[:-1])
-        try:
-            parent.__setattr__(path[-1], value)
-        except KeyError:
-            raise KeyError("'%s'" % key)
-
-class MetaNodeIterator(six.Iterator):
-    """
-    An iterator for the meta node which flattens the hierachical structure
-    """
-    def __init__(self, node):
-        self.key_stack = []
-        self.iter_stack = [six.iteritems(node._instance)]
-    
-    def __iter__(self):
-        return self
-    
-    def __next__(self):
-        while self.iter_stack:
-            try:
-                key, val = six.next(self.iter_stack[-1])
-            except StopIteration:
-                self.iter_stack.pop()
-                if self.iter_stack:
-                    self.key_stack.pop()
-                continue
-                
-            if isinstance(val, dict):
-                self.key_stack.append(key)
-                self.iter_stack.append(six.iteritems(val))
-            else:
-                return '.'.join(self.key_stack + [key])
-                
-        raise StopIteration
